+++ CMakeLists.txt
@@ -12,4 +12,5 @@ add_executable(Game_tDND
         src/Item.cpp
         src/Person.cpp
         Include/Untill.h
+        src/Function.cpp
 )
