+++ Include/Inventory.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <vector>
 #include "Item.h"
