+++ src/Game0.2.cpp
@@ -31,10 +31,12 @@ int main()
         std::cout << "Player selection failed!" << std::endl;
         return 1; 
     }
+    Player->inventory.AddItem(bow);
+    Player->inventory.AddItem(sword);
+    Player->inventory.AddItem(healthPotion);
     InventoryRealization(*Player);
     cin.get();
-    Player->inventory.AddItem(bow);
-    Player->EquipWeapon(1);
+    Player->EquipError(1);
     std::cout << "Starting Battle..." << std::endl;
     Battle(*Player);
